ft/service security
