+++ api-gateway/src/main/java/com/api_gateway/api_gateway/config/WebClientConfig.java
@@ -1,0 +1,14 @@
+package com.api_gateway.api_gateway.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+    @Bean
+    public WebClient.Builder webClientBuilder() {
+        return WebClient.builder();
+    }
+}
