+++ api-gateway/src/main/java/com/api_gateway/api_gateway/dto/ValidateTokenResponseDto.java
@@ -1,0 +1,15 @@
+package com.api_gateway.api_gateway.dto;
+
+public class ValidateTokenResponseDto {
+    private String email;
+
+    // Getters and setters
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+}
