+++ api-gateway/src/main/java/com/api_gateway/api_gateway/exception/UnauthorizedException.java
@@ -1,0 +1,12 @@
+package com.api_gateway.api_gateway.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+// This will return a 401 Unauthorized response status
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
