+++ api-gateway/src/main/resources/application.properties
@@ -4,3 +4,7 @@ spring.main.web-application-type=reactive
 management.endpoints.web.exposure.include=*
 management.endpoint.health.show-details=always
 eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka/
+
+security.jwt.token.secret-key=${JWT_SECRET_KEY}
+security.jwt.token.expiration=${DB_JWT_EXPIRATION}
+auth.service.url=http://localhost:8093/users
