+++ api-gateway/src/main/resources/application.yml
@@ -1,45 +1,82 @@
-# Application Configuration
 spring:
   application:
     name: api-gateway
 
-  # Cloud Gateway Configuration
   cloud:
     gateway:
-      httpclient:
-        connect-timeout: 20000
-        response-timeout: 7s
+      discovery:
+        locator:
+          enabled: true
+          lower-case-service-id: true
       routes:
-        # Route to Trainee Management Service
+        # Route to Trainee Management Service (Protected)
         - id: trainee-management-service
           uri: lb://trainee-management-service
           predicates:
             - Path=/trainees/**
+          filters:
+            - name: AuthFilter
 
         # Route to User Management Service
-        - id: user-management-service
+        - id: user-management-service-login
           uri: lb://user-management-service
           predicates:
-            - Path=/users/**
+            - Path=/users/login/**, /users/{id}/password
 
-        # Route to Email Management Service
+        - id: user-management-service-create
+          uri: lb://user-management-service
+          predicates:
+            - Path=/users/create
+          filters:
+            - name: AuthFilter
+
+        # Route to Email Management Service (Protected)
         - id: email-management-service
-          uri: lb://email
+          uri: lb://email-management-service
           predicates:
             - Path=/email/**
 
+        # Route to Authentication Service
+        - id: authentication-service
+          uri: lb://authentication-service
+          predicates:
+            - Path=/auth/login, /auth/generate-otp/**, /auth/verify-otp/**, /auth/register, /auth/validate-token/**, /departments/**, /roles/**, /user-roles/**, /auth/reset-password/**
+          filters:
+            - name: AuthFilter
 
-  # Eureka Client Configuration
+        # Route for creating passwords
+        - id: authentication-service-create-password
+          uri: lb://authentication-service
+          predicates:
+            - Path=/auth/create-password, /users/**
+          filters:
+            - name: AuthFilter
+
+        # Route for SOP Management Service (Protected)
+        - id: sop-management-service
+          uri: lb://sop-management-service
+          predicates:
+            - Path=/sop-management/**
+          filters:
+            - name: AuthFilter
+
+        # Route for Email Service
+        - id: email-service
+          uri: lb://email-service
+          predicates:
+            - Path=/email-service/**
+
+  # Eureka Configuration
   eureka:
     client:
       service-url:
         defaultZone: http://localhost:8761/eureka/
-      instance:
-        prefer-ip-address: true
-        lease-renewal-interval-in-seconds: 30
-        lease-expiration-duration-in-seconds: 90
+    instance:
+      prefer-ip-address: true
+      lease-renewal-interval-in-seconds: 30
+      lease-expiration-duration-in-seconds: 90
 
-# Server Configuration
+# Server Port
 server:
   port: 8093
 
@@ -48,7 +85,7 @@ logging:
   level:
     org.springframework.cloud.gateway: DEBUG
 
-# Swagger API Documentation Configuration
+# Swagger API Documentation
 springdoc:
   enable-native-support: true
   api-docs:
