--- email-management-service/src/main/java/com/springemail/email/UserRepository.java
@@ -1,15 +1,0 @@
-package com.springemail.email;
-
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByResetToken(String token);
-
-    // If you are using a reset token
-
-}
