--- kafka-consumer/src/main/java/com/kafka_consumer/kafka_consumer/KafkaConsumerApplication.java
@@ -1,13 +1,0 @@
-package com.kafka_consumer.kafka_consumer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KafkaConsumerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KafkaConsumerApplication.class, args);
-	}
-
-}
