--- kafka-consumer/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.application.name=kafka-consumer
-server.port=8095
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
-spring.main.web-application-type=reactive
-spring.kafka.bootstrap-servers=localhost:9092
-
-spring.kafka.consumer.group-id=group_id
