--- kafka-consumer/src/main/resources/application.yml
@@ -1,28 +1,0 @@
-# Configuration for Kafka Consumer
-spring:
-  config:
-    activate:
-      on-profile: kafka-consumer
-
-  application:
-    name: kafka-consumer
-
-server:
-  port: 8095  # Kafka Consumer listens on port 8095
-
-eureka:
-  client:
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
-    instance:
-      prefer-ip-address: true
-      lease-renewal-interval-in-seconds: 30
-      lease-expiration-duration-in-seconds: 90
-
-kafka:
-  bootstrap-servers: localhost:9092  # Point to your Kafka broker
-  consumer:
-    group-id: consumer-group  # Set a consumer group ID
-    key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-    value-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-    max-partition-fetch-bytes: 200000000  # 200 MB
