--- kafka-consumer/src/test/java/com/kafka_consumer/kafka_consumer/KafkaConsumerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kafka_consumer.kafka_consumer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KafkaConsumerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
