--- kafka-producer/src/main/java/com/kafka_producer/kafka_producer/config/KafkaTopicConfig.java
@@ -1,15 +1,0 @@
-package com.kafka_producer.kafka_producer.config;
-
-import org.apache.kafka.clients.admin.NewTopic;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
-
-@Configuration
-public class KafkaTopicConfig {
-    @Bean
-    public NewTopic topic1() {
-return TopicBuilder.name("topic1")
-        .build();
-    }
-}
