--- kafka-producer/src/main/java/com/kafka_producer/kafka_producer/dto/MessageRequest.java
@@ -1,12 +1,0 @@
-package com.kafka_producer.kafka_producer.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageRequest {
-    private String message;
-}
