--- kafka-producer/src/main/java/com/kafka_producer/kafka_producer/service/MessageProducer.java
@@ -1,20 +1,0 @@
-package com.kafka_producer.kafka_producer.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MessageProducer {
-
-    private final KafkaTemplate<String, String> kafkaTemplate;
-
-    @Autowired
-    public MessageProducer(KafkaTemplate<String, String> kafkaTemplate) {
-        this.kafkaTemplate = kafkaTemplate;
-    }
-
-    public void sendMessage(String topic, String message) {
-        kafkaTemplate.send(topic, message);
-    }
-}
