--- kafka-producer/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.application.name=kafka-producer
-server.port=8094
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
-spring.main.web-application-type=reactive
-spring.kafka.bootstrap-servers=localhost:9092
