--- kafka-producer/src/main/resources/application.yml
@@ -1,27 +1,0 @@
-# Configuration for Kafka Producer
-spring:
-  config:
-    activate:
-      on-profile: kafka-producer
-
-  application:
-    name: kafka-producer
-
-server:
-  port: 8094  # Kafka Producer listens on port 8094
-
-eureka:
-  client:
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
-    instance:
-      prefer-ip-address: true
-      lease-renewal-interval-in-seconds: 30
-      lease-expiration-duration-in-seconds: 90
-
-kafka:
-  bootstrap-servers: localhost:9092  # Point to your Kafka broker
-  producer:
-    key-serializer: org.apache.kafka.common.serialization.StringSerializer
-    value-serializer: org.apache.kafka.common.serialization.StringSerializer
-    max-partition-fetch-bytes: 200000000  # 200 MB
