--- kafka-producer/src/test/java/com/kafka_producer/kafka_producer/KafkaProducerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kafka_producer.kafka_producer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KafkaProducerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
