--- user-management-service/k8s/k8s/user-management-deployment.yml
@@ -1,32 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: user-management-service
-spec:
-  replicas: 3
-  selector:
-    matchLabels:
-      app: user-management
-  template:
-    metadata:
-      labels:
-        app: user-management
-    spec:
-      containers:
-        - name: user-management
-          image: your-docker-repo/user-management-service:latest
-          ports:
-            - containerPort: 8083
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: user-management-service
-spec:
-  selector:
-    app: user-management
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 8083
-  type: LoadBalancer
