+++ user-management-service/src/main/java/com/user_management/user_management_service/dto/MessageRequest.java
@@ -1,0 +1,12 @@
+package com.user_management.user_management_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class MessageRequest {
+    private String message;
+}
