+++ user-management-service/src/main/resources/application.properties
@@ -40,5 +40,6 @@ email.service.url=http://localhost:8093/email
 logging.level.org.springframework.security=DEBUG
 kafka.topic.user-created=user-created
 spring.kafka.bootstrap-servers=localhost:9092
+spring.kafka.consumer.group-id=email-service
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.StringDeserializer
